+++ system/modules/core/library/Contao/Controller.php
@@ -92,18 +92,23 @@ abstract class Controller extends \System
 			$arrTemplates[$strTemplate] = $strTemplate;
 		}
 
-		// Add the customized templates
-		foreach (glob(TL_ROOT . '/templates/' . $strPrefix . '*') as $strFile)
-		{
-			$strTemplate = basename($strFile, strrchr($strFile, '.'));
+		$arrCustomized = glob(TL_ROOT . '/templates/' . $strPrefix . '*');
 
-			if (!isset($arrTemplates[$strTemplate]))
+		// Add the customized templates
+		if (is_array($arrCustomized))
+		{
+			foreach ($arrCustomized as $strFile)
 			{
-				$arrTemplates[$strTemplate] = $strTemplate;
+				$strTemplate = basename($strFile, strrchr($strFile, '.'));
+
+				if (!isset($arrTemplates[$strTemplate]))
+				{
+					$arrTemplates[$strTemplate] = $strTemplate;
+				}
 			}
 		}
 
-		// Wrap into a try-catch block (see #5210)
+		// Try to select the themes (see #5210)
 		try
 		{
 			$objTheme = \ThemeModel::findAll(array('order'=>'name'));
@@ -120,13 +125,18 @@ abstract class Controller extends \System
 			{
 				if ($objTheme->templates != '')
 				{
-					foreach (glob(TL_ROOT . '/' . $objTheme->templates . '/' . $strPrefix . '*') as $strFile)
-					{
-						$strTemplate = basename($strFile, strrchr($strFile, '.'));
+					$arrThemeTemplates = glob(TL_ROOT . '/' . $objTheme->templates . '/' . $strPrefix . '*');
 
-						if (!isset($arrTemplates[$strTemplate]))
+					if (is_array($arrThemeTemplates))
+					{
+						foreach ($arrThemeTemplates as $strFile)
 						{
-							$arrTemplates[$strTemplate] = $strTemplate . ' (' . sprintf($GLOBALS['TL_LANG']['MSC']['templatesTheme'], $objTheme->name) . ')';
+							$strTemplate = basename($strFile, strrchr($strFile, '.'));
+
+							if (!isset($arrTemplates[$strTemplate]))
+							{
+								$arrTemplates[$strTemplate] = $strTemplate . ' (' . sprintf($GLOBALS['TL_LANG']['MSC']['templatesTheme'], $objTheme->name) . ')';
+							}
 						}
 					}
 				}
@@ -1740,6 +1750,12 @@ abstract class Controller extends \System
 	 */
 	public static function generateMargin($arrValues, $strType='margin')
 	{
+		// Initialize an empty array (see #5217)
+		if (!is_array($arrValues))
+		{
+			$arrValues = array('top'=>'', 'right'=>'', 'bottom'=>'', 'left'=>'', 'unit'=>'');
+		}
+
 		$top = $arrValues['top'];
 		$right = $arrValues['right'];
 		$bottom = $arrValues['bottom'];
@@ -1766,15 +1782,8 @@ abstract class Controller extends \System
 			}
 		}
 
-		$arrDir = array
-		(
-			'top'=>$top,
-			'right'=>$right,
-			'bottom'=>$bottom,
-			'left'=>$left
-		);
-
 		$return = array();
+		$arrDir = array('top'=>$top, 'right'=>$right, 'bottom'=>$bottom, 'left'=>$left);
 
 		foreach ($arrDir as $k=>$v)
 		{